% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins_rle}
\alias{margins_rle}
\alias{margins_dense}
\title{Calculate row and column counts, sums, and sums-of-squares}
\usage{
margins_rle(fname, group, bufsize = 1e+07, n = Inf)

margins_dense(fname, group, bufsize = 6000, n = Inf)
}
\arguments{
\item{fname}{character(1) HDF5 file name to 10xGenomics-format file.}

\item{group}{character(1) group name (e.g., `mm10`) containing counts.}

\item{bufsize}{integer(1) approximate maximum number of non-zero
counts (for `margins_rle()`) or columns (for `margins_dense()`)
to read per iteration. The default is reasonably memory- and
time-efficient.}

\item{n}{numeric(1) number of blocks to process; default: all.}
}
\value{
a `list()` with `row` and `column` elements. Each element
    is itself a list with `n`, `sum`, and `sumsq` (sum-of-squares)
    of counts in the corresponding row or column.
}
\description{
Calculate row and column counts, sums, and sums-of-squares
}
\details{
Use `margins_rle()` to process 10xGenomics 'run-length
    encoded' HDF5 files.

Use `margins_dense()` to process _Bioconductor_'s
    rectangular (dense matrix) HDF5 files.
}
