% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins}
\alias{margins}
\title{Calculate row and column counts, sums, and sums-of-squares}
\usage{
margins(fname, group, bufsize = 1e+07, n = Inf)
}
\arguments{
\item{fname}{character(1) HDF5 file name to 10xGenomics-format file.}

\item{group}{character(1) group name (e.g., `mm10`) containing counts.}

\item{bufsize}{integer(1) approximate maximum number of counts to
read per iteration. The default is reasonably memory- and
time-efficient.}

\item{n}{numeric(1) number of iterations to perform; default: all.}
}
\value{
a `list()` with `row` and `column` elements. Each element
    is itself a list with `n`, `sum`, and `sumsq` (sum-of-squares)
    of counts in the corresponding row or column.
}
\description{
Calculate row and column counts, sums, and sums-of-squares
}
